plugin_proxy: enable event_type specification for proxy plugins
This commit adds event_type to the flb_plugin_proxy_def struct. This allows for the setting
of event_type to a value other than the default log only initialization.

In the flb_plugin_proxy_register function default behaviour is enforced by checking for unset event_type vlaues.
When unset (a value of 0) then log behaviour is defaulted. No need for incorrect value checking since this
was not a set field in the past. Unset fields are set as current behavior dicates and future usecases using incorrect values can be treated as user error.

Signed-off-by: jmccormick7 <johnmccormick03711@gmail.com>